[usage] Remove unused WorkspaceInstanceUsage model
